radix sort fixed digit count complete
the radix sort on a list of NON-NEGATIVE numbers with same number of digits is implemented and
working. the sot takes in a vector in the constructor, sorts its internal list when the sort is
called, then outputs a sorted vector when called to do so. Unit tests have been added to the
driver.